Fix Convex on Ethereum
