+++ src/adapters/convex-finance/ethereum/index.ts
@@ -1,8 +1,10 @@
-import { getConvexPoolsContracts } from '@adapters/convex-finance/ethereum/_pool'
 import { getConvexGaugesBalances } from '@adapters/convex-finance/ethereum/balance'
+import { getConvexPoolsContracts } from '@adapters/convex-finance/ethereum/pool'
+import { getCvxCrvStakeBalance, getCVXStakeBalance } from '@adapters/convex-finance/ethereum/stake'
 import { getPoolsContracts } from '@adapters/curve-dex/ethereum/pools'
-import type { BaseContext, Contract, GetBalancesHandler } from '@lib/adapter'
+import type { BalancesContext, BaseContext, Contract, GetBalancesHandler } from '@lib/adapter'
 import { resolveBalances } from '@lib/balance'
+import { getMultipleLockerBalances } from '@lib/lock'
 import type { Token } from '@lib/token'
 
 const threeCrv: Token = {
@@ -85,24 +87,32 @@ const cvxCRVStaker: Contract = {
 export const getContracts = async (ctx: BaseContext) => {
   const curvePools = await getPoolsContracts(ctx, metaRegistry)
   const pools = await getConvexPoolsContracts(ctx, booster, curvePools)
-  // const pools = await getPoolsContracts(ctx, booster)
 
   return {
     contracts: {
-      // cvxCRVStaker,
-      // cvxRewardPool,
-      // locker,
+      cvxCRVStaker,
+      cvxRewardPool,
+      locker,
       pools,
     },
   }
 }
 
+async function getConvexBalances(ctx: BalancesContext, pools: Contract[]) {
+  const [gaugeBalances, crvRewardsBalances] = await Promise.all([
+    getConvexGaugesBalances(ctx, pools, metaRegistry, 'gauge'),
+    getConvexGaugesBalances(ctx, pools, metaRegistry, 'crvRewards'),
+  ])
+
+  return [...(gaugeBalances ?? []), ...(crvRewardsBalances ?? [])]
+}
+
 export const getBalances: GetBalancesHandler<typeof getContracts> = async (ctx, contracts) => {
   const balances = await resolveBalances<typeof getContracts>(ctx, contracts, {
-    pools: (...args) => getConvexGaugesBalances(...args, metaRegistry),
-    // cvxRewardPool: getCVXStakeBalance,
-    // cvxCRVStaker: getCvxCrvStakeBalance,
-    // locker: (...args) => getMultipleLockerBalances(...args, CVX, [cvxCRV, cvxFXS, FXS], true),
+    pools: getConvexBalances,
+    cvxRewardPool: getCVXStakeBalance,
+    cvxCRVStaker: getCvxCrvStakeBalance,
+    locker: (...args) => getMultipleLockerBalances(...args, CVX, [cvxCRV, cvxFXS, FXS], true),
   })
 
   return {
